+++ netbox_custom_scripts/tables/__init__.py
@@ -1,10 +1,11 @@
 from .module import CustomScriptModuleTable
-from .project import CustomScriptProjectRevisionTable, CustomScriptProjectTable
+from .project import CustomScriptProjectFileTable, CustomScriptProjectRevisionTable, CustomScriptProjectTable
 from .script import CustomScriptLogTable, CustomScriptTable
 
 __all__ = (
     'CustomScriptLogTable',
     'CustomScriptModuleTable',
+    'CustomScriptProjectFileTable',
     'CustomScriptProjectRevisionTable',
     'CustomScriptProjectTable',
     'CustomScriptTable',